Move the PlayingBar tick logic to avoid useless Queue rendering
